+++ orm-patterns-instances/CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - Examples for Hibernate/spring data JPA.
 
+## 2020-11-09
+
+### Added
+- The update method for phone in hibernate-jpa examples project.
+
 ## 2020-11-08
 
 ### Added